move most of main to cmd
